+++ crates/gmsol/src/store/events.rs
@@ -1,9 +1,10 @@
 use crate::{
     decode::untagged,
     types::{
-        DepositExecuted, DepositRemoved, GlvDepositRemoved, GlvWithdrawalRemoved, GtUpdated,
-        MarketFeesUpdated, MarketStateUpdated, OrderRemoved, PositionDecreased, PositionIncreased,
-        ShiftRemoved, SwapExecuted, TradeEvent, WithdrawalExecuted, WithdrawalRemoved,
+        BorrowingFeesUpdated, DepositExecuted, DepositRemoved, GlvDepositRemoved,
+        GlvWithdrawalRemoved, GtUpdated, MarketFeesUpdated, MarketStateUpdated, OrderRemoved,
+        PositionDecreased, PositionIncreased, ShiftRemoved, SwapExecuted, TradeEvent,
+        WithdrawalExecuted, WithdrawalRemoved,
     },
 };
 
@@ -22,6 +23,7 @@ untagged!(
         OrderRemoved,
         TradeEvent,
         MarketFeesUpdated,
+        BorrowingFeesUpdated,
         MarketStateUpdated,
         SwapExecuted,
         GtUpdated